+++ test/test_group.py
@@ -9,7 +9,10 @@ from draftsman.classes.blueprint import Blueprint
 from draftsman.classes.collision_set import CollisionSet
 from draftsman.classes.entity_list import EntityList
 from draftsman.classes.group import Group
+from draftsman.classes.schedule import Schedule, WaitCondition, WaitConditions
+from draftsman.classes.schedule_list import ScheduleList
 from draftsman.classes.vector import Vector
+from draftsman.constants import Direction, WaitConditionType
 from draftsman.entity import *
 from draftsman.error import (
     DraftsmanError,
@@ -90,6 +93,35 @@ class GroupTesting(unittest.TestCase):
         )
         assert group.entities[2].neighbours[0]() is group.entities[3]
         assert group.entities[3].neighbours[0]() is group.entities[2]
+
+        # Initialize from blueprint string with entities and schedules
+        test_string_with_schedules = "0eNqVk91uwjAMhd/F1xmiLf2928WeYkIotB5Ya5MqSdkQ6rvPIWxFiImi3NSNz+c4OT7Bth2wN6QcVCegWisL1fsJLO2UbP0/d+wRKiCHHQhQsvORkdTCKIBUg99QRaN4KGl1rTvt6IBXwnhcC0DlyBGGwufguFFDt0XD5Ht6Ab22LNHKV2PMS5SUizSLUwFHqMqcC2hDjJIhabnIU3/EG3j8B7eOO9rt3cu5sTv8VTKhGzJYh934DjV5ghrNpq7mU5NyNjV9gprPpmZPUNN/qOwLW++xGdqLMab393F0tR8M+/vW8PqmHBr2tPVO+ZLkNmzr5lw3oBjUS4Obi0+14bzLt6POqxzVn5wbFcul7+9GIFUzKUhJPvSB+590yWPZx9DyBK39EmFSqqtZFHBAY8NVFNEqL+M8y9IiK4px/AGk8DmW"
+        group = Group(name="test", position=(2, 2), string=test_string_with_schedules)
+        assert group.name == "test"
+        assert group.type == "group"
+        assert group.id == None
+        assert group.position == Vector(2.0, 2.0)
+        assert len(group.entities) == 6
+        assert group.entities == EntityList(initlist=[
+            Locomotive("locomotive", position=(-137.5625, 99.0), orientation=0.75),
+            StraightRail("straight-rail", position=(-141, 99), direction=Direction.EAST),
+            StraightRail("straight-rail", position=(-139, 99), direction=Direction.EAST),
+            StraightRail("straight-rail", position=(-137, 99), direction=Direction.EAST),
+            StraightRail("straight-rail", position=(-135, 99), direction=Direction.EAST),
+            StraightRail("straight-rail", position=(-133, 99), direction=Direction.EAST),
+        ])
+        assert len(group.schedules) == 1
+        assert group.schedules[0].locomotives[0]() is group.entities[0]
+        assert group.schedules[0].stops == [
+            {
+                "station": "AEnterprise",
+                "wait_conditions": WaitConditions([
+                    WaitCondition(type="time", compare_type="or", ticks=1800),
+                    WaitCondition(type="inactivity", compare_type="and"),
+                    WaitCondition(type="full", compare_type="and")
+                ])
+            }
+        ]
 
         # Initialize from blueprint string with no entities
         empty_blueprint_string = "0eNqrVkrKKU0tKMrMK1GyqlbKLEnNVbJCEtNRKkstKs7Mz1OyMrIwNDG3NDI3NTUxMTU3q60FAHmbE1Y="
@@ -325,6 +357,40 @@ class GroupTesting(unittest.TestCase):
                 "1": {"red": [{"entity_id": Association(group.entities[0])}]}
             },
         }
+
+    def test_set_schedules(self):
+        # Regular list
+        group = Group()
+        group.schedules = []
+        assert isinstance(group.schedules, ScheduleList)
+
+        group.entities.append("locomotive", id="test_train")
+
+        # ScheduleList
+        schedule = Schedule()
+        schedule.add_locomotive(group.entities["test_train"])
+        schedule.append_stop("station_name", WaitCondition(WaitConditionType.INACTIVITY, ticks=600))
+        group.schedules = ScheduleList([schedule])
+        assert isinstance(group.schedules, ScheduleList)
+        assert group.schedules[0].locomotives[0]() is group.entities[0]
+        assert group.schedules[0].stops == [
+            {
+                "station": "station_name",
+                "wait_conditions": WaitConditions([WaitCondition(WaitConditionType.INACTIVITY, ticks=600)])
+            }
+        ]
+
+        # None
+        group.schedules = None
+        assert isinstance(group.schedules, ScheduleList)
+        assert group.schedules == ScheduleList()
+        assert len(group.schedules) == 0
+
+        with pytest.raises(TypeError):
+            group.schedules = dict()
+
+        with pytest.raises(TypeError):
+            group.schedules = ["incorrect", "format"]
 
     def test_power_connections(self):
         group = Group("test")
@@ -849,6 +915,12 @@ class GroupTesting(unittest.TestCase):
         group.add_circuit_connection("red", 0, 1)
         group.entities.append("power-switch", tile_position=(1, 1))
         group.add_power_connection(2, 3, side=1)
+        group.entities.append("locomotive", id="loco", tile_position = (0, 0))
+        schedule = Schedule()
+        schedule.add_locomotive(group.entities["loco"])
+        schedule.append_stop("ore", WaitCondition("full"))
+        schedule.append_stop("dropoff", WaitCondition("empty"))
+        group.schedules = [schedule]
 
         blueprint = Blueprint()
         blueprint.entities.append(group, copy=False)
@@ -868,6 +940,10 @@ class GroupTesting(unittest.TestCase):
         )
         assert (
             group.entities[3].connections["Cu0"][0]["entity_id"]() is group.entities[2]
+        )
+        # Make sure schedule associations are preserved
+        assert (
+            group.schedules[0].locomotives[0]() is group.entities["loco"]
         )
         # Make sure the parent is correct
         assert group.parent is blueprint
@@ -892,6 +968,10 @@ class GroupTesting(unittest.TestCase):
         assert (
             group_copy.entities[3].connections["Cu0"][0]["entity_id"]()
             is group_copy.entities[2]
+        )
+        # Make sure schedule associations are preserved
+        assert (
+            group_copy.schedules[0].locomotives[0]() is group_copy.entities["loco"]
         )
         # Make sure parent of the copied group is reset to None
         assert group_copy.parent is None